Feature/article rest api

+++ django-back/articles/views.py
@@ -3,8 +3,8 @@ from rest_framework.decorators import api_view, permission_classes
 from rest_framework.response import Response
 from rest_framework.permissions import IsAuthenticated
 from .models import Article, Comment
-from .serializers import ArticleListSerializer, ArticleSerializer, CommentSerializer
-from .permissions import IsAuthorOrReadonly
+from .serializers import ArticleListSerializer, ArticleSerializer
+from .serializers import CommentSerializer, CommentListSerializer
 
 
 # Create your views here.
@@ -15,6 +15,7 @@ def article_list(request):
     return Response(serializer.data)
 
 @api_view(['GET'])
+@permission_classes([IsAuthenticated])
 def article_detail(request, article_pk):
     article = get_object_or_404(Article, pk=article_pk)
     serializer = ArticleSerializer(article)
@@ -29,14 +30,46 @@ def article_create(request):
         return Response(serializer.data)
 
 @api_view(['PUT', 'DELETE'])
-@permission_classes([IsAuthorOrReadonly])
 def article_update(request, article_pk):
     article = get_object_or_404(Article, pk=article_pk)
-    if request.method == 'PUT':
-        serializer = ArticleSerializer(data=request.data, instance=article)
-        if serializer.is_valid(raise_exception=True):
-            serializer.save()
-            return Response(serializer.data)
+    if request.user.id == article.user.id:
+        if request.method == 'PUT':
+            serializer = ArticleSerializer(data=request.data, instance=article)
+            if serializer.is_valid(raise_exception=True):
+                serializer.save()
+                return Response(serializer.data)
+        else:
+            article.delete()
+            return Response({'message':'Article has been deleted!'})
     else:
-        article.delete()
-        return Response({'message':'Article has been deleted!'})
+        return Response({'message':'author only'})
+
+@api_view(['GET'])
+def comment_list(request, article_pk):
+    comments = Comment.objects.filter(article=article_pk)
+    serializer = CommentListSerializer(comments, many=True)
+    return Response(serializer.data)
+
+@api_view(['POST'])
+@permission_classes([IsAuthenticated])
+def comment_create(request, article_pk):
+    article = get_object_or_404(Article, pk=article_pk)
+    serializer = CommentSerializer(data=request.data)
+    if serializer.is_valid(raise_exception=True):
+        serializer.save(user=request.user, article=article)
+        return Response(serializer.data)
+
+@api_view(['PUT', 'DELETE'])
+def comment_update(request, article_pk, comment_pk):
+    comment = get_object_or_404(Comment, pk=comment_pk)
+    if request.user.id == comment.user.id:
+        if request.method == 'PUT':
+            serializer = CommentSerializer(data=request.data, instance=comment)
+            if serializer.is_valid(raise_exception=True):
+                serializer.save()
+                return Response(serializer.data)
+        else:
+            comment.delete()
+            return Response({'message':'Comment has been deleted!'})
+    else:
+        return Response({'message':'author only'})
